requirements.txt
flet
pygame
opencv-python
easyocr
numpy
Pillow
openpyxl